Move example tree sequences to tsutil
