+++ python/tests/test_balance_metrics.py
@@ -29,7 +29,7 @@ import pytest
 
 import tests
 import tskit
-from tests.test_highlevel import get_example_tree_sequences
+from tests.tsutil import get_example_tree_sequences
 
 # ↑ See https://github.com/tskit-dev/tskit/issues/1804 for when
 # we can remove this.
