+++ python/tests/test_extend_haplotypes.py
@@ -6,7 +6,7 @@ import _tskit
 import tests.test_wright_fisher as wf
 import tskit
 from tests import tsutil
-from tests.test_highlevel import get_example_tree_sequences
+from tests.tsutil import get_example_tree_sequences
 
 # ↑ See https://github.com/tskit-dev/tskit/issues/1804 for when
 # we can remove this.
