+++ python/tests/test_ibd.py
@@ -10,7 +10,7 @@ import tests
 import tests.ibd as ibd
 import tests.test_wright_fisher as wf
 import tskit
-from tests.test_highlevel import get_example_tree_sequences
+from tests.tsutil import get_example_tree_sequences
 
 """
 Tests of IBD finding algorithms.
