+++ python/tests/test_phylo_formats.py
@@ -36,7 +36,7 @@ from Bio import SeqIO
 
 import tests
 import tskit
-from tests.test_highlevel import get_example_tree_sequences
+from tests.tsutil import get_example_tree_sequences
 
 # ↑ See https://github.com/tskit-dev/tskit/issues/1804 for when
 # we can remove this. The example_ts here is intended to be the
