+++ python/tests/test_table_transforms.py
@@ -33,7 +33,7 @@ import pytest
 import tests
 import tskit
 import tskit.util as util
-from tests.test_highlevel import get_example_tree_sequences
+from tests.tsutil import get_example_tree_sequences
 
 # ↑ See https://github.com/tskit-dev/tskit/issues/1804 for when
 # we can remove this.
